CurrencyLogo refactor
