+++ src/components/CurrencyLogo/index.ts
@@ -1,0 +1,4 @@
+export { CurrencyLogo } from './CurrencyLogo.container'
+export { CurrencyWrapperSource } from './CurrencyLogo.types'
+// eslint-disable-next-line
+export type { CurrencyLogoContainerProps } from './CurrencyLogo.types'
