+++ src/components/LiquidityMining/Create/Steps/PairAndReward/AssetSelector/index.tsx
@@ -2,7 +2,7 @@ import { Token } from '@swapr/sdk'
 import React, { useEffect, useState } from 'react'
 import styled from 'styled-components'
 import { TYPE } from '../../../../../../theme'
-import CurrencyLogo from '../../../../../CurrencyLogo'
+import { CurrencyLogo } from '../../../../../CurrencyLogo'
 import DoubleCurrencyLogo from '../../../../../DoubleLogo'
 import { Box, Flex } from 'rebass'
 import { GradientCard } from '../../../../../Card'
