+++ src/components/LiquidityMining/Create/Steps/RewardAmount/index.tsx
@@ -3,7 +3,7 @@ import React, { useCallback, useEffect, useState } from 'react'
 import { Box, Flex } from 'rebass'
 import styled from 'styled-components'
 import { TYPE } from '../../../../../theme'
-import CurrencyLogo from '../../../../CurrencyLogo'
+import { CurrencyLogo } from '../../../../CurrencyLogo'
 import NumericalInput from '../../../../Input/NumericalInput'
 import Radio from '../../../../Radio'
 import { Card, Divider } from '../../../styleds'
