+++ src/components/Pool/LiquidityMiningCampaignView/Information/index.tsx
@@ -12,7 +12,7 @@ import { ExternalLink, TYPE } from '../../../../theme'
 import { CarrotButton } from '../../../Button'
 import { AutoColumn } from '../../../Column'
 import Countdown from '../../../Countdown'
-import CurrencyLogo from '../../../CurrencyLogo'
+import { CurrencyLogo } from '../../../CurrencyLogo'
 import DoubleCurrencyLogo from '../../../DoubleLogo'
 import Row, { AutoRow, RowBetween, RowFixed } from '../../../Row'
 import DataDisplayer from '../DataDisplayer'
