+++ src/components/Pool/LiquidityMiningCampaignView/TokenAmountDisplayer/index.tsx
@@ -3,7 +3,7 @@ import React, { useContext } from 'react'
 import { Box, Flex } from 'rebass'
 import { useNativeCurrencyUSDPrice } from '../../../../hooks/useNativeCurrencyUSDPrice'
 import { TYPE } from '../../../../theme'
-import CurrencyLogo from '../../../CurrencyLogo'
+import { CurrencyLogo } from '../../../CurrencyLogo'
 import { MouseoverTooltip } from '../../../Tooltip'
 import { AutoColumn } from '../../../Column'
 import { AutoRow } from '../../../Row'
