+++ src/components/SearchModalBridge/CommonTokens.tsx
@@ -5,7 +5,7 @@ import { ChainId, Currency, currencyEquals, Token } from '@swapr/sdk'
 import { SUGGESTED_BASES } from '../../constants'
 import { AutoColumn } from '../Column'
 import { AutoRow } from '../Row'
-import CurrencyLogo from '../CurrencyLogo'
+import { CurrencyLogo } from '../CurrencyLogo'
 import { TYPE } from '../../theme'
 import { useNativeCurrency } from '../../hooks/useNativeCurrency'
 import styled from 'styled-components'
