+++ src/components/SearchModalBridge/CurrencyList.tsx
@@ -6,7 +6,7 @@ import { useActiveWeb3React } from '../../hooks'
 import { useAddUserToken, useRemoveUserAddedToken } from '../../state/user/hooks'
 import { useCurrencyBalances } from '../../state/wallet/hooks'
 import { useIsUserAddedToken } from '../../hooks/Tokens'
-import CurrencyLogo from '../CurrencyLogo'
+import { CurrencyLogo } from '../CurrencyLogo'
 import Loader from '../Loader'
 import Badge from '../Badge'
 import { TokenPickerItem } from './styleds'
