+++ src/components/SearchModalBridge/ImportRow.tsx
@@ -2,7 +2,7 @@ import React, { CSSProperties, useContext } from 'react'
 import { Token } from '@swapr/sdk'
 import { AutoRow, RowFixed } from '../Row'
 import { AutoColumn } from '../Column'
-import CurrencyLogo from '../CurrencyLogo'
+import { CurrencyLogo } from '../CurrencyLogo'
 import { TYPE } from '../../theme'
 import ListLogo from '../ListLogo'
 import { ButtonPrimary } from '../Button'
