+++ src/components/SearchModalBridge/ManageTokens.tsx
@@ -7,7 +7,7 @@ import { useToken } from '../../hooks/Tokens'
 import styled, { ThemeContext } from 'styled-components'
 import { useUserAddedTokens, useRemoveUserAddedToken } from '../../state/user/hooks'
 import { Token } from '@swapr/sdk'
-import CurrencyLogo from '../CurrencyLogo'
+import { CurrencyLogo } from '../CurrencyLogo'
 import { isAddress } from '../../utils'
 import { useActiveWeb3React } from '../../hooks'
 import Card from '../Card'
