+++ src/components/TokenWarningModal/index.tsx
@@ -4,7 +4,7 @@ import styled, { ThemeContext } from 'styled-components'
 import { useActiveWeb3React } from '../../hooks'
 import { ExternalLink, TYPE } from '../../theme'
 import { getExplorerLink, shortenAddress } from '../../utils'
-import CurrencyLogo from '../CurrencyLogo'
+import { CurrencyLogo } from '../CurrencyLogo'
 import Modal from '../Modal'
 import { AutoRow, RowFixed } from '../Row'
 import { AutoColumn } from '../Column'
