+++ src/components/swap/SwapPlatformSelector.tsx
@@ -2,7 +2,7 @@ import React, { useCallback } from 'react'
 import { CurrencyAmount, Percent, Trade, RoutablePlatform, TradeType, UniswapV2Trade } from '@swapr/sdk'
 import { AutoColumn } from '../Column'
 import { TYPE } from '../../theme'
-import CurrencyLogo from '../CurrencyLogo'
+import { CurrencyLogo } from '../CurrencyLogo'
 import { Box, Flex } from 'rebass'
 import Radio from '../Radio'
 import QuestionHelper from '../QuestionHelper'
